NIT: Reuse javaVersion method - remove dot as github includes it into the clickable URL - Move methods for downloading to Update.scala - apply fmt
